requirements.txt
board
busio
adafruit_ht16k33.segments
digitalio
adafruit_mcp230xx.mcp23017